requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.1.0
fastapi==0.108.0
frozenlist==1.4.1
h11==0.14.0
idna==3.7
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
starlette==0.32.0.post1
typing_extensions==4.12.2
upstash-redis==1.0.0
urllib3==2.2.2
uvicorn==0.25.0
yarl==1.9.4
